Separate decorator tests
